WA-1742: Add config option to set default characters, add migration function, break off plugin into smaller files.
